+++ eims/src/main/java/com/example/eims/dto/report/IncomeReportItemDTO.java
@@ -2,15 +2,17 @@ package com.example.eims.dto.report;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class IncomeReportItemDTO {
-    private Float total;
-    private Float paid;
+    private BigDecimal total;
+    private BigDecimal paid;
 
     public IncomeReportItemDTO() {
     }
 
-    public IncomeReportItemDTO(Float total, Float paid) {
+    public IncomeReportItemDTO(BigDecimal total, BigDecimal paid) {
         this.total = total;
         this.paid = paid;
     }
